Move Target to base
