+++ omni-cx2x/src/cx2x/Cx2x.java
@@ -9,7 +9,7 @@ import cx2x.decompiler.XcmlBackend;
 import cx2x.translator.ClawTranslatorDriver;
 import cx2x.translator.config.Configuration;
 import cx2x.translator.language.accelerator.AcceleratorDirective;
-import cx2x.translator.language.helper.target.Target;
+import cx2x.translator.language.base.Target;
 import cx2x.translator.report.ClawTransformationReport;
 import exc.xcodeml.XcodeMLtools_Fmod;
 import org.apache.commons.cli.*;
