+++ omni-cx2x/src/cx2x/translator/common/Utility.java
@@ -6,7 +6,7 @@
 package cx2x.translator.common;
 
 import cx2x.translator.language.accelerator.AcceleratorDirective;
-import cx2x.translator.language.helper.target.Target;
+import cx2x.translator.language.base.Target;
 import xcodeml.util.XmOption;
 
 import java.util.*;
