+++ omni-cx2x/src/cx2x/translator/language/accelerator/generator/OpenMp.java
@@ -8,7 +8,7 @@ package cx2x.translator.language.accelerator.generator;
 import cx2x.translator.common.Utility;
 import cx2x.translator.config.Configuration;
 import cx2x.translator.language.accelerator.AcceleratorDirective;
-import cx2x.translator.language.helper.target.Target;
+import cx2x.translator.language.base.Target;
 
 import java.util.List;
