+++ omni-cx2x/src/cx2x/translator/language/base/ClawLanguage.java
@@ -11,7 +11,6 @@ import cx2x.translator.language.common.ClawRange;
 import cx2x.translator.language.common.ClawReshapeInfo;
 import cx2x.translator.language.accelerator.AcceleratorDirective;
 import cx2x.translator.language.accelerator.generator.AcceleratorGenerator;
-import cx2x.translator.language.helper.target.Target;
 import cx2x.translator.language.parser.ClawLexer;
 import cx2x.translator.language.parser.ClawParser;
 import cx2x.xcodeml.exception.IllegalDirectiveException;
