+++ omni-cx2x/src/cx2x/translator/language/parser/Claw.g4
@@ -20,7 +20,7 @@ grammar Claw;
 import cx2x.translator.common.ClawConstant;
 import cx2x.translator.language.base.*;
 import cx2x.translator.language.common.*;
-import cx2x.translator.language.helper.target.Target;
+import cx2x.translator.language.base.Target;
 import cx2x.translator.common.Utility;
 import cx2x.xcodeml.language.*;
 }
