+++ omni-cx2x/unittest/cx2x/translator/language/ClawEnumTest.java
@@ -7,7 +7,7 @@ package cx2x.translator.language;
 import cx2x.translator.common.ClawConstant;
 import cx2x.translator.language.common.OverPosition;
 import cx2x.translator.language.accelerator.AcceleratorDirective;
-import cx2x.translator.language.helper.target.Target;
+import cx2x.translator.language.base.Target;
 import cx2x.translator.xnode.ClawAttr;
 import org.junit.Test;
